refactor: opt code
