+++ packages/loader/src/gltf/extensions/Galacean_materials_remap.ts
@@ -1,4 +1,4 @@
-import { Material } from "@oasis-engine/core";
+import { Material } from "@galacean/engine-core";
 import { registerGLTFExtension } from "../parser/GLTFParser";
 import { GLTFParserContext } from "../parser/GLTFParserContext";
 import { GLTFExtensionMode, GLTFExtensionParser } from "./GLTFExtensionParser";
