+++ packages/loader/src/gltf/extensions/KHR_materials_ior.ts
@@ -1,13 +1,12 @@
 import { PBRMaterial } from "@galacean/engine-core";
-import { IKHRMaterialsIor } from "./Schema";
+import { IKHRMaterialsIor } from "./GLTFExtensionSchema";
+import { GLTFParserContext } from "../parser";
+import { registerGLTFExtension } from "../parser/GLTFParser";
+import { GLTFExtensionMode, GLTFExtensionParser } from "./GLTFExtensionParser";
 
-import { registerExtension } from "../parser/Parser";
-import { ParserContext } from "../parser/ParserContext";
-import { ExtensionParser } from "./ExtensionParser";
-
-@registerExtension("KHR_materials_ior")
-class KHR_materials_ior extends ExtensionParser {
-  parseEngineResource(schema: IKHRMaterialsIor, material: PBRMaterial, context: ParserContext): void {
+@registerGLTFExtension("KHR_materials_ior", GLTFExtensionMode.AdditiveParse)
+class KHR_materials_ior extends GLTFExtensionParser {
+  parseEngineResource(schema: IKHRMaterialsIor, material: PBRMaterial, context: GLTFParserContext): void {
     const { ior = 1.5 } = schema;
 
     material.ior = ior;
