+++ packages/loader/src/gltf/extensions/KHR_materials_variants.ts
@@ -1,4 +1,4 @@
-import { Renderer，Material } from "@galacean/engine-core";
+import { Material, Renderer } from "@galacean/engine-core";
 import { registerGLTFExtension } from "../parser/GLTFParser";
 import { GLTFParserContext } from "../parser/GLTFParserContext";
 import { GLTFExtensionMode, GLTFExtensionParser } from "./GLTFExtensionParser";
